deployment/requirements.txt
streamlit
requests
numpy
tensorflow-cpu
pillow